refactor: Long -> Integer 및 entity 수정

+++ airdnb/src/main/java/team10/airdnb/accommodation/controller/request/AccommodationCreateRequest.java
@@ -27,7 +27,7 @@ public record AccommodationCreateRequest(
         String name,
 
         @Min(1) @Max(16)
-        Long maxCapacity,
+        Integer maxCapacity,
 
         Long accommodationType,
